requirements.txt
django==3.1.5
django-crispy-forms==1.10.0
Pillow==8.1.0
djangorestframework==3.12.2
matplotlib
opencv-python
numpy
djangorestframework-simplejwt
djoser
djangorestframework-jwt